🔧 Move command (handler) to own directory
Based on the aggregate the command is for
